Modifications for seeding and implementation of foreign keys TO DO : debug & seeding fake data for transitory tables

+++ api/totallyNotLastFm/database/migrations/2018_04_28_110307_create_compose_table.php
@@ -15,8 +15,8 @@ class CreateComposeTable extends Migration
   {
     Schema::create('compose', function (Blueprint $table) {
       $table->engine = 'InnoDB';
-      $table->increments('artist_id_artist', 255);
-      $table->integer('music_id_music');
+      $table->unsignedBigInteger('artist_id_artist', 255);
+      $table->unsignedBigInteger('music_id_music');
       $table->nullableTimestamps();
     });
   }
